# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: AIX — Artificial Intelligence eXperience API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /user/profile:
    get:
      operationId: getUserProfile
      tags: [user]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: upsertUserProfile
      tags: [user]
      summary: Create or update user profile (registration)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileInput"
      responses:
        "200":
          description: Profile saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  # ─── SUBJECTS ─────────────────────────────────────────────────────────
  /subjects:
    get:
      operationId: listSubjects
      tags: [subjects]
      summary: List all subjects
      responses:
        "200":
          description: List of subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"
    post:
      operationId: createSubject
      tags: [subjects]
      summary: Create a subject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectInput"
      responses:
        "201":
          description: Subject created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
  /subjects/{subjectId}:
    get:
      operationId: getSubject
      tags: [subjects]
      summary: Get a subject by ID
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateSubject
      tags: [subjects]
      summary: Update a subject
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectUpdate"
      responses:
        "200":
          description: Subject updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
    delete:
      operationId: deleteSubject
      tags: [subjects]
      summary: Delete a subject
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
  /subjects/{subjectId}/progress:
    get:
      operationId: getSubjectProgress
      tags: [subjects]
      summary: Get progress stats for a subject
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Progress stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubjectProgress"

  # ─── LESSONS ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    DeleteResult:
      type: object
      properties:
        success:
          type: boolean
      required: [success]

    # ─── USER ─────────────────────────────────────────────────────────
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        age:
          type: integer
        educationLevel:
          type: string
        school:
          type: string
        major:
          type: ["string", "null"]
        country:
          type: string
        timezone:
          type: string
        annualGoal:
          type: string
        targetGrade:
          type: number
        dailyStudyHours:
          type: number
        sleepTime:
          type: string
        wakeTime:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - age
        - educationLevel
        - school
        - country
        - timezone
        - annualGoal
        - targetGrade
        - dailyStudyHours
        - sleepTime
        - wakeTime
        - createdAt
        - updatedAt
    UserProfileInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        age:
          type: integer
          minimum: 5
          maximum: 100
        educationLevel:
          type: string
          minLength: 1
        school:
          type: string
          minLength: 1
        major:
          type: ["string", "null"]
        country:
          type: string
          minLength: 1
        timezone:
          type: string
        annualGoal:
          type: string
          minLength: 1
        targetGrade:
          type: number
          minimum: 0
          maximum: 100
        dailyStudyHours:
          type: number
          minimum: 0
          maximum: 24
        sleepTime:
          type: string
        wakeTime:
          type: string
      required:
        - name
        - age
        - educationLevel
        - school
        - country
        - timezone
        - annualGoal
        - targetGrade
        - dailyStudyHours
        - sleepTime
        - wakeTime

    # ─── SUBJECTS ───────────────────────────────────────────────────
    Subject:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        teacher:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, name, color, difficulty, createdAt]
    SubjectInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        color:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        teacher:
          type: ["string", "null"]
      required: [name, color, difficulty]
    SubjectUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        color:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        teacher:
          type: ["string", "null"]
    SubjectProgress:
      type: object
      properties:
        subjectId:
          type: integer
        totalLessons:
          type: integer
        completedLessons:
          type: integer
        reviewedLessons:
          type: integer
        averageMastery:
          type: number
        completionRate:
          type: number
      required:
        - subjectId
        - totalLessons
        - completedLessons
        - reviewedLessons
        - averageMastery
        - completionRate

    # ─── LESSONS ─────────────────────────────────────────────────────
...