# lib/api-zod/src/generated/types/subjectUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * AIX — Artificial Intelligence eXperience API
 * OpenAPI spec version: 1.0.0
 */
import type { SubjectUpdateDifficulty } from './subjectUpdateDifficulty';

export interface SubjectUpdate {
  /** @minLength 1 */
  name?: string;
  color?: string;
  difficulty?: SubjectUpdateDifficulty;
  /** @nullable */
  teacher?: string | null;
}